+++ src/main/java/Formes/FormeGeometrique.java
@@ -10,5 +10,5 @@ public abstract class  FormeGeometrique  {
 	/*
 	 * deplacement d'une figure en abcisse et ordonnées
 	 */
-	public abstract void move(double dx, double dy);
+	public abstract void move(int dx, int dy);
 }
